change types of mutation-ids to `usize`, was: `u32`

+++ mutagen-core/src/comm/coverage.rs
@@ -2,5 +2,5 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct CoverageHit {
-    pub mutator_id: u32,
+    pub mutator_id: usize,
 }
